# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance media conversion API delivering WhatsApp-ready audio and images.\n\n**Author:** Guilherme Jansen · [GitHub](https://github.com/guilhermejansen)\n**Contato direto:** suporte@setupautomatizado.com.br\n**Coleção Postman pronta:** [Download](./swagger/postman.json)\n\nUse os exemplos documentados em cada endpoint para testar rapidamente conversões de áudio, imagem e operações S3.",
        "title": "WhatsApp Media Converter API",
        "contact": {
            "name": "Guilherme Jansen",
            "url": "https://github.com/guilhermejansen",
            "email": "suporte@setupautomatizado.com.br"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/convert/audio": {
            "post": {
                "description": "Accepts base64 payloads or multipart uploads and returns an optimized Opus data URI.",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Conversion"
                ],
                "summary": "Convert audio to WhatsApp-compatible Opus format",
                "parameters": [
                    {
                        "description": "Audio conversion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_services.AudioRequest"
                        }
                    },
                    {
                        "type": "file",
                        "description": "Audio file when using multipart",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_services.AudioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/convert/batch/audio": {
            "post": {
                "description": "Processes up to 10 audio conversion jobs concurrently.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Conversion"
                ],
                "summary": "Convert a batch of audio payloads",
                "parameters": [
                    {
                        "description": "Batch audio conversion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/whats-convert-api_internal_services.AudioRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.BatchAudioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/convert/batch/image": {
            "post": {
                "description": "Processes up to 10 image conversion jobs concurrently.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Conversion"
                ],
                "summary": "Convert a batch of image payloads",
                "parameters": [
                    {
                        "description": "Batch image conversion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/whats-convert-api_internal_services.ImageRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.BatchImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/convert/image": {
            "post": {
                "description": "Accepts base64 payloads or multipart uploads and returns a compressed JPEG data URI.",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Conversion"
                ],
                "summary": "Convert image to WhatsApp-optimized JPEG",
                "parameters": [
                    {
                        "description": "Image conversion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_services.ImageRequest"
                        }
                    },
                    {
                        "type": "file",
                        "description": "Image file when using multipart",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_services.ImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/whats-convert-api_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "whats-convert-api_internal_models.BatchAudioResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 2
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/whats-convert-api_internal_services.AudioResponse"
                    }
                }
            }
        },
        "whats-convert-api_internal_models.BatchImageResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 2
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/whats-convert-api_internal_services.ImageResponse"
                    }
                }
            }
        },
...
        "whats-convert-api_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "Missing 'data' field"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
        "whats-convert-api_internal_services.AudioRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "base64 or URL",
                    "type": "string",
                    "example": "data:audio/aac;base64,T2dnUwACAAAAAAAAAAB"
                },
                "input_type": {
                    "description": "Optional: mp3, wav, m4a, etc.",
                    "type": "string",
                    "example": "mp3"
                },
                "is_url": {
                    "description": "true if data is URL",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "whats-convert-api_internal_services.AudioResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "base64 opus audio",
                    "type": "string",
                    "example": "data:audio/ogg;codecs=opus;base64,T2dnUwACAAAA"
                },
                "duration": {
                    "description": "Duration in seconds",
                    "type": "integer",
                    "example": 8
                },
                "size": {
                    "description": "Size in bytes",
                    "type": "integer",
                    "example": 42144
                }
            }
        },
        "whats-convert-api_internal_services.ImageRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "base64 or URL",
                    "type": "string",
                    "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD"
                },
                "is_url": {
                    "description": "true if data is URL",
                    "type": "boolean",
                    "example": false
                },
                "max_height": {
                    "description": "Optional: max height (default 1920)",
                    "type": "integer",
                    "example": 1920
                },
                "max_width": {
                    "description": "Optional: max width (default 1920)",
                    "type": "integer",
                    "example": 1920
                },
                "quality": {
                    "description": "Optional: JPEG quality 1-100 (default 95)",
                    "type": "integer",
                    "example": 90
                }
            }
        },
        "whats-convert-api_internal_services.ImageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "base64 jpeg image",
                    "type": "string",
                    "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABA"
                },
                "height": {
                    "description": "Image height",
                    "type": "integer",
                    "example": 600
                },
                "size": {
                    "description": "Size in bytes",
                    "type": "integer",
                    "example": 20480
                },
                "width": {
                    "description": "Image width",
                    "type": "integer",
                    "example": 800
                }
            }
        }
    }
}